check param before cleaning local storage

+++ components/Success.jsx
@@ -1,23 +1,34 @@
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 import Link from "next/link";
 import { BsBagCheckFill } from "react-icons/bs";
 import { useStateContext } from "@/context/StateContext";
 import { runFireworks } from "@/lib/utils";
 import style from "../styles/Success.module.css";
+import { useSearchParams } from "next/navigation";
 
 const { successWrapper, success, icon, emailMsg, description, email } = style;
 
 const Success = () => {
   const { setCartItems, setTotalPrice, setTotalQuantities } = useStateContext();
-  useEffect(() => {
-    localStorage.clear();
-    setCartItems([]);
-    setTotalPrice(0);
-    setTotalQuantities(0);
-    runFireworks();
-  }, []);
+  const searchParams = useSearchParams();
 
-  return (
+  const successCallback = searchParams.get("success");
+
+  console.log(successCallback);
+
+  useEffect(() => {
+    if (successCallback) {
+      console.log("useeffect");
+      // updateInventory
+      localStorage.clear();
+      setCartItems([]);
+      setTotalPrice(0);
+      setTotalQuantities(0);
+      runFireworks();
+    }
+  }, [successCallback]);
+
+  return success ? (
     <div className={successWrapper}>
       <div className={success}>
         <BsBagCheckFill className={icon} />
@@ -37,6 +48,8 @@ const Success = () => {
         </Link>
       </div>
     </div>
+  ) : (
+    <p>Il n'y a rien à voir ici.</p>
   );
 };
